requirements.txt
# asgiref>=3.4.1
# Django>=3.2.13
# django-cors-headers>=3.7.0
# django-extensions>=3.1.3
# django-ninja>=0.13.2
# djangorestframework>=3.12.4
# ninja>=1.10.0.post2
# pydantic>=1.8.2
# pytz>=2021.1
# sqlparse>=0.4.1
# typing-extensions>=3.10.0.0
asgiref==3.4.1
Django==3.2.5
django-cors-headers==3.7.0
django-extensions==3.1.3
django-ninja==0.13.2
djangorestframework==3.12.4
ninja==1.10.0.post2
pydantic==1.8.2
pytz==2021.1
sqlparse==0.4.1
typing-extensions==3.10.0.0